// BookingTicketsUser/src/app/Models/invoiceShipping.model .ts
export class InvoiceShipping {
    id : number;
    recipientName : string;
    recipientPhone : string;
    recipientAddress : string;
    deliveryName : string;
    deliveryPhone: string;
    deliveryAddress: string;
    idAccount: number;
    idShipping:number;
    createAt: string;
    updateAt: string;
    status: boolean;
    payment: string;
    total: number;
}